code tidying VIII

+++ src/web/api_handlers_timeline.c
@@ -271,14 +271,14 @@ void handle_get_timeline_segments(const http_request_t *req, http_response_t *re
         if (segments[i].size_bytes < 1024) {
             snprintf(size_str, sizeof(size_str), "%ld B", segments[i].size_bytes);
         } else if (segments[i].size_bytes < (uint64_t)1024 * 1024) {
-            snprintf(size_str, sizeof(size_str), "%.1f KB", segments[i].size_bytes / 1024.0);
+            snprintf(size_str, sizeof(size_str), "%.1f KB", (double)segments[i].size_bytes / 1024.0);
         } else if (segments[i].size_bytes < (uint64_t)1024 * 1024 * 1024) {
-            snprintf(size_str, sizeof(size_str), "%.1f MB", segments[i].size_bytes / (1024.0 * 1024.0));
+            snprintf(size_str, sizeof(size_str), "%.1f MB", (double)segments[i].size_bytes / (1024.0 * 1024.0));
         } else {
-            snprintf(size_str, sizeof(size_str), "%.1f GB", segments[i].size_bytes / (1024.0 * 1024.0 * 1024.0));
+            snprintf(size_str, sizeof(size_str), "%.1f GB", (double)segments[i].size_bytes / (1024.0 * 1024.0 * 1024.0));
         }
-        
-        cJSON_AddNumberToObject(segment, "id", segments[i].id);
+
+        cJSON_AddNumberToObject(segment, "id", (double)segments[i].id);
         cJSON_AddStringToObject(segment, "stream", segments[i].stream_name);
         cJSON_AddStringToObject(segment, "start_time", segment_start_time);
         cJSON_AddStringToObject(segment, "end_time", segment_end_time);
