+++ src/web/api_handlers_users_backend_agnostic.c
@@ -26,7 +26,7 @@
 static cJSON *user_to_json(const user_t *user, int include_api_key) {
     cJSON *json = cJSON_CreateObject();
 
-    cJSON_AddNumberToObject(json, "id", user->id);
+    cJSON_AddNumberToObject(json, "id", (double)user->id);
     cJSON_AddStringToObject(json, "username", user->username);
     cJSON_AddStringToObject(json, "email", user->email);
     cJSON_AddNumberToObject(json, "role", user->role);
@@ -36,9 +36,9 @@ static cJSON *user_to_json(const user_t *user, int include_api_key) {
         cJSON_AddStringToObject(json, "api_key", user->api_key);
     }
 
-    cJSON_AddNumberToObject(json, "created_at", user->created_at);
-    cJSON_AddNumberToObject(json, "updated_at", user->updated_at);
-    cJSON_AddNumberToObject(json, "last_login", user->last_login);
+    cJSON_AddNumberToObject(json, "created_at", (double)user->created_at);
+    cJSON_AddNumberToObject(json, "updated_at", (double)user->updated_at);
+    cJSON_AddNumberToObject(json, "last_login", (double)user->last_login);
     cJSON_AddBoolToObject(json, "is_active", user->is_active);
     cJSON_AddBoolToObject(json, "password_change_locked", user->password_change_locked);
     cJSON_AddBoolToObject(json, "totp_enabled", user->totp_enabled);
